data/integrations/requirements.txt
# Dependencias del framework de sincronización
# =============================================

# Base de datos
psycopg2-binary>=2.9.0

# HTTP requests
requests>=2.28.0

# Google Sheets API
google-api-python-client>=2.0.0
google-auth>=2.0.0
google-auth-oauthlib>=0.5.0
google-auth-httplib2>=0.1.0

# Retry logic
tenacity>=8.0.0

# Circuit breaker (opcional)
circuitbreaker>=1.0.0

# Caching (opcional)
cachetools>=5.0.0

# Webhooks
flask>=2.0.0

# Scheduler
croniter>=1.3.0

# Testing
pytest>=7.0.0
pytest-cov>=4.0.0

# OCR and Document Processing
# ===========================

# Tesseract OCR (local)
pytesseract>=0.3.10
Pillow>=10.0.0

# Google Cloud Vision API
google-cloud-vision>=3.4.0

# Azure Computer Vision (via requests, pero también puedes usar SDK)
# azure-cognitiveservices-vision-computervision>=1.0.0

# AWS Textract (opcional)
# boto3>=1.26.0

# OpenAI Vision API (opcional)
# openai>=1.0.0

# PDF processing
PyPDF2>=3.0.0
pdf2image>=1.16.0  # Requiere poppler-utils instalado en el sistema

# Image processing
opencv-python>=4.8.0
numpy>=1.24.0
scipy>=1.10.0

# Flask for API
flask>=2.3.0
flask-cors>=4.0.0

# Queue processing
redis>=4.5.0

# Cloud storage
boto3>=1.26.0  # AWS S3
google-cloud-storage>=2.10.0  # Google Cloud Storage

# Drive integrations
google-api-python-client>=2.0.0
google-auth-oauthlib>=0.5.0
dropbox>=11.36.0

# Image processing for signatures
scikit-image>=0.21.0

# Advanced export formats
openpyxl>=3.1.0  # Excel export
reportlab>=4.0.0  # PDF reports

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-mock>=3.11.0

# Machine Learning
scikit-learn>=1.3.0
sentence-transformers>=2.2.0

# PDF compression
PyMuPDF>=1.23.0

# Additional utilities
python-dateutil>=2.8.0

# Translation
googletrans==4.0.0-rc1  # Google Translate
deepl>=1.15.0  # DeepL (opcional, mejor calidad)

# Security
cryptography>=41.0.0
PyJWT>=2.8.0

# GraphQL
graphql-core>=3.2.0

# NLP and NER
spacy>=3.7.0
# python -m spacy download es_core_news_sm

# Summarization
transformers>=4.30.0  # Para resumen abstractivo (opcional)
torch>=2.0.0  # Requerido por transformers

# Vector databases
pinecone-client>=2.2.0  # Pinecone
qdrant-client>=1.6.0  # Qdrant

# Sentiment analysis
vaderSentiment>=3.3.2
textblob>=0.17.1

# Search indexing
elasticsearch>=8.10.0  # Para indexación y búsqueda avanzada
